CatchClauseMarker has been renamed to the ExceptionHandlerMarker.
Stack simulator: Simulate the stack operation and for each instruction construct the status of stack, which contains the instructions when a stack operand is pushed onto the stack. For instance, take 'instr_a' into consideration, we are storing all the possible locations, where the first operand on the stack is pushed, from the view of 'instr_a', and then the second operand, third and so on. 

<Idea from Lukas>