BasicBlockStaticContext: drop the per-method CFG cache keyed by class, method name, and method descriptor.
This causes problems when the same class is loaded multiple times by
different class loaders. The basic block static context will use CFG
that was made for a different method and will not find instruction
nodes of the new method, leading to all sorts of errors.

Instead, memoize the last CFG along with the corresponding method
node, and if the method node changes, recreate the CFG. This seems
to be enough to avoid regenerating the CFG for the same method.

To some degree, this is the result of poor static context interface
and reuse of a single static context instance for multiple shadows.
This needs to be redesigned some time in future.