Call shutdown() on socket before close() in DiSL and Shadow VM agents
This is to ensure that data that may still be in socket
buffers when close() is called are sent to the other side.
This mainly applies to the Shadow VM agent, because it only
sends data to the Shadow VM server and does not expect any
reply, but it is probably a good practice in general.