Remove setMultipleValDelim() from the marker Parameter class
This was a source of issues with markers which forgot to set the
delimiter prior to requesting multiple values. Instead, the
getMultipleValues() method has been modified to require the delimiter
as an argument, forcing the users of the Parameter class to supply the
delimiter on use. The BytecodeMarker, InsnNodeMarker, and the
StrictBytecodeMarker were updated to reflect the change.